updated licensing
As discussed at the end of our meeting, we want the majority of the code to be open and have minimal restrictions so that we and others may use it in the future on different projects. At the same time, we want our game to be closed, so that we have all options open to us later when we decide what to do with it.